refactor: 틀린 파일 이름을 수정
- announcementDone -> applyDone
- announcementDoneTemp -> applyDoneTemp
