+++ service-apply/src/pages/apply/ApplyDoneTemp.page.tsx
@@ -1,7 +1,7 @@
 import { Button, Txt } from '@quokka/design-system';
 import { Link } from 'react-router-dom';
 
-export const AnnouncementDoneTempPage = () => {
+export const ApplyDoneTempPage = () => {
   return (
     <div className="h-[100dvh] flex justify-center items-center flex-col gap-8 text-center">
       <Txt size="h2">임시 저장이 완료 되었습니다.</Txt>
