+++ service-apply/src/router/index.tsx
@@ -2,24 +2,21 @@ import { Routes, Route } from 'react-router-dom';
 import { PasswordResetPage } from '../pages/PasswordReset/PasswordReset.page';
 import { RequestPasswordResetPage } from '../pages/PasswordReset/RequestPasswordReset.page';
 import { HomePage } from '../pages/Home.page';
-import { ApplyPage } from '../pages/Apply.page';
+import { ApplyPage } from '../pages/apply/Apply.page';
 import { AnnouncementListPage } from '../pages/Anouncement/AnnouncementList.page';
 import { AnnouncementPage } from '../pages/Anouncement/Announcement.page';
 import { AnnouncementLayoutPage } from '../pages/Anouncement/AnounceLayout.page';
 import { PasswordResetLayout } from '../pages/PasswordReset/PasswordResetLayout.page';
-import { AnnouncementDonePage } from '../pages/Anouncement/AnnouncementDone.page';
-import { AnnouncementDoneTempPage } from '../pages/Anouncement/AnnouncementDoneTemp.page';
+import { ApplyDonePage } from '../pages/apply/ApplyDone.page';
+import { ApplyDoneTempPage } from '../pages/apply/ApplyDoneTemp.page';
 
 export function Router() {
   return (
     <Routes>
       <Route path="/" element={<HomePage />} />
       <Route path="/apply" element={<ApplyPage />} />
-      <Route path="/announcement/done" element={<AnnouncementDonePage />} />
-      <Route
-        path="/announcement/done/temp"
-        element={<AnnouncementDoneTempPage />}
-      />
+      <Route path="/announcement/done" element={<ApplyDonePage />} />
+      <Route path="/announcement/done/temp" element={<ApplyDoneTempPage />} />
       <Route path="/announcement" element={<AnnouncementLayoutPage />}>
         <Route index element={<AnnouncementListPage />} />
         <Route path=":announcementId" element={<AnnouncementPage />} />
